+++ src/app/static/src/app/store/baseline/actions.ts
@@ -3,9 +3,8 @@ import {BaselineState} from "./baseline";
 import {api} from "../../apiService";
 import {PjnzResponse, PopulationResponse, ShapeResponse, ValidateBaselineResponse} from "../../generated";
 import {BaselineMutation} from "./mutations";
-import qs from "qs";
-import {findResource, getFilenameFromImportUrl, getFilenameFromUploadFormData} from "../../utils";
-import { DatasetResourceSet, DatasetResource, ADRSchemas } from "../../types";
+import {buildData, findResource, getFilenameFromImportUrl, getFilenameFromUploadFormData} from "../../utils";
+import {DatasetResourceSet, DatasetResource, ADRSchemas, UploadImportPayload} from "../../types";
 import {DataExplorationState} from "../dataExploration/dataExploration";
 import {initialChorplethSelections} from "../plottingSelections/plottingSelections";
 
@@ -34,7 +33,7 @@ const uploadCallback = async (dispatch: Dispatch, response: any) => {
 
 interface UploadImportOptions {
     url: string
-    payload: FormData | string
+    payload: FormData | UploadImportPayload
 }
 
 
@@ -97,7 +96,7 @@ async function uploadOrImportShape(context: ActionContext<BaselineState, DataExp
 export const actions: ActionTree<BaselineState, DataExplorationState> & BaselineActions = {
 
     async refreshDatasetMetadata(context) {
-        const { commit, state, rootState, rootGetters } = context
+        const { commit, state, rootState, rootGetters, dispatch } = context
         if (state.selectedDataset) {
             let url = `/adr/datasets/${state.selectedDataset.id}`;
             if (state.selectedDataset.release) {
@@ -111,18 +110,18 @@ export const actions: ActionTree<BaselineState, DataExplorationState> & Baseline
                 .then((response) => {
                     if (response) {
                         const metadata = response.data;
-                        const avaliableResources = rootGetters ? rootGetters["baseline/selectedDatasetAvailableResources"] : {}
+                        const availableResources = rootGetters ? rootGetters["baseline/selectedDatasetAvailableResources"] : {}
                         const exceptions = { // where DatasetResource keys do not match ADRSchemas keys
                             pop: "population",
                             program: "programme"
                         }
                         const resources: { [k in keyof DatasetResourceSet]?: DatasetResource | null } = {}
 
-                        Object.entries(avaliableResources).forEach(([key, value]) => {
+                        Object.entries(availableResources).forEach(([key, value]) => {
                             // If an available resource has a value, find the resource (using the alternate schema key where needed)
                             // and add it to the committed object under the original key
                             const schemaKey = key in exceptions ? exceptions[key as "pop" || "program"] : key
-                            resources[key as keyof typeof resources] = value ? findResource(metadata, schemas[schemaKey as keyof ADRSchemas]) : null
+                            resources[key as keyof typeof resources] = key ? findResource(metadata, schemas[schemaKey as keyof ADRSchemas]) : null
                         })
 
                         commit(BaselineMutation.UpdateDatasetResources,
@@ -134,21 +133,24 @@ export const actions: ActionTree<BaselineState, DataExplorationState> & Baseline
 
     async importPJNZ(context, url) {
         if (url) {
-            await uploadOrImportPJNZ(context, {url: "/adr/pjnz/", payload: qs.stringify({url})},
+            const data = buildData (context.state.selectedDataset, url, "pjnz")
+            await uploadOrImportPJNZ(context, {url: "/adr/pjnz/", payload: data},
                 getFilenameFromImportUrl(url));
         }
     },
 
     async importPopulation(context, url) {
         if (url) {
-            await uploadOrImportPopulation(context, {url: "/adr/population/", payload: qs.stringify({url})},
+            const data = buildData (context.state.selectedDataset, url, "pop")
+            await uploadOrImportPopulation(context, {url: "/adr/population/", payload: data},
                 getFilenameFromImportUrl(url));
         }
     },
 
     async importShape(context, url) {
         if (url) {
-            await uploadOrImportShape(context, {url: "/adr/shape/", payload: qs.stringify({url})},
+            const data = buildData (context.state.selectedDataset, url, "shape")
+            await uploadOrImportShape(context, {url: "/adr/shape/", payload: data},
                 getFilenameFromImportUrl(url));
         }
     },