+++ src/app/static/src/app/types.ts
@@ -317,3 +317,22 @@ interface DownloadIds {
     summary: string,
     coarse_output: string
 }
+
+export interface UploadImportPayload {
+    url: string,
+    datasetId: string,
+    resourceId: string
+}
+
+export type DatasetResourceType = "pjnz" | "pop"  | "program" | "anc" | "shape" | "survey";
+
+export interface DownloadIndicatorDataset {
+    filteredData: unknown[],
+    unfilteredData: unknown[]
+}
+
+export interface DownloadIndicatorPayload {
+    data: DownloadIndicatorDataset,
+    filename: string
+}
+