+++ NEWS.md
@@ -1,3 +1,19 @@
+# hint 2.32.1
+
+* As a user I can see ADR source URL in output zip for input data so I know the source of the files
+
+# hint 2.32.0
+
+* As a user I can download input time series data
+
+# hint 2.31.3
+
+* Styling of ADR key links
+
+# hint 2.31.2
+
+* Make refresh button realtime
+
 # hint 2.31.1
 
 * Setup Playwright and add Browser test for data exploration login